requirements.txt
# Azure SDK
azure-cosmos==4.5.1
azure-functions==1.18.0
azure-eventhub==5.11.4
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-monitor-opentelemetry==1.2.0
azure-ai-ml==1.13.0
azure-communication-email==1.0.0
azure-communication-sms==1.0.1

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.3
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
imbalanced-learn==0.12.0

# Deep Learning (sequence model)
torch>=2.1.0

# Model Management & Tracking
mlflow>=2.10.0
shap==0.44.1

# Data Processing
pyarrow==15.0.0
openpyxl==3.1.2

# Visualization
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pydantic==2.5.3
tenacity==8.2.3

# Logging & Monitoring
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# Date/Time
python-dateutil==2.8.2
pytz>=2020.1

# HTTP & APIs
requests==2.31.0
aiohttp==3.9.3
fredapi==0.5.2

# Kafka Streaming
confluent-kafka==2.3.0
kafka-python==2.0.2
